// src/libs/data-access/klass/models/CorrespondenceTableResource.ts
export interface CorrespondenceTableResource {
    /**
     * 
     * @type {string}
     * @memberof CorrespondenceTableResource
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof CorrespondenceTableResource
     */
    id?: number;
    /**
     * 
     * @type {ContactPersonResource}
     * @memberof CorrespondenceTableResource
     */
    contactPerson?: ContactPersonResource;
    /**
     * 
     * @type {string}
     * @memberof CorrespondenceTableResource
     */
    owningSection?: string;
    /**
     * 
     * @type {string}
     * @memberof CorrespondenceTableResource
     */
    source?: string;
    /**
     * 
     * @type {number}
     * @memberof CorrespondenceTableResource
     */
    sourceId?: number;
    /**
     * 
     * @type {string}
     * @memberof CorrespondenceTableResource
     */
    target?: string;
    /**
     * 
     * @type {number}
     * @memberof CorrespondenceTableResource
     */
    targetId?: number;
    /**
     * 
     * @type {boolean}
     * @memberof CorrespondenceTableResource
     */
    changeTable?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof CorrespondenceTableResource
     */
    lastModified?: Date;
    /**
     * 
     * @type {Array<string>}
     * @memberof CorrespondenceTableResource
     */
    published?: Array<string>;
    /**
     * 
     * @type {LevelResource}
     * @memberof CorrespondenceTableResource
     */
    sourceLevel?: LevelResource;
    /**
     * 
     * @type {LevelResource}
     * @memberof CorrespondenceTableResource
     */
    targetLevel?: LevelResource;
    /**
     * 
     * @type {string}
     * @memberof CorrespondenceTableResource
     */
    description?: string;
    /**
     * 
     * @type {Array<ChangelogResource>}
     * @memberof CorrespondenceTableResource
     */
    changelogs?: Array<ChangelogResource>;
    /**
     * 
     * @type {Array<CorrespondenceMapResource>}
     * @memberof CorrespondenceTableResource
     */
    correspondenceMaps?: Array<CorrespondenceMapResource>;
    /**
     * 
     * @type {{ [key: string]: Link; }}
     * @memberof CorrespondenceTableResource
     */
    links?: { [key: string]: Link; };
    /**
     * 
     * @type {ClassificationVariantSummaryResourceEmbedded}
     * @memberof CorrespondenceTableResource
     */
    embedded?: ClassificationVariantSummaryResourceEmbedded;
    /**
     * 
     * @type {PageMetadata}
     * @memberof CorrespondenceTableResource
     */
    page?: PageMetadata;
}